increase project rules coverage
