+++ test/integration_tests/python/conftest.py
@@ -197,6 +197,16 @@ def log(request):
     return log
 
 
+@pytest.fixture(scope='function')
+def with_user(data_builder, randstr, as_public):
+    """Return AttrDict with new user, api-key and api-accessor"""
+    api_key = randstr()
+    user = data_builder.create_user(api_key=api_key, root=True)
+    session = copy.deepcopy(as_public)
+    session.headers.update({'Authorization': 'scitran-user ' + api_key})
+    return attrdict.AttrDict(user=user, api_key=api_key, session=session)
+
+
 class BaseUrlSession(requests.Session):
     """Requests session subclass using core api's base url"""
     def request(self, method, url, **kwargs):
